requirements.txt
kaggle==1.6.14
pandas==2.2.0
boto3==1.34.0
python-dotenv==1.0.0
pyarrow==15.0.0
